+++ changelog_entry.yaml
@@ -1,0 +1,6 @@
+- bump: minor
+  changes:
+    added:
+    - Colorado family affordability tax credit.
+    removed:
+    - Draft version of Colorado family affordability tax credit.
