--- policyengine_us/parameters/gov/contrib/co_general_assembly/family_affordability_credit/reduction/older/start.yaml
@@ -1,20 +1,0 @@
-description: The Colorado General Assembly reduces the family affordability tax credit starting at this adjusted gross income threshold, for older children, based on filing status.
-metadata:
-  period: year
-  unit: currency-USD
-  label: Colorado General Assembly family affordability tax credit older reduction start
-  breakdown:
-    - filing_status
-  reference:
-    - title: Colorado General Assembly House Bill 24-1311 Family Affordability Tax Credit
-      href: https://leg.colorado.gov/bills/hb24-1311
-SINGLE:
-  2015-01-01: 15_000
-HEAD_OF_HOUSEHOLD:
-  2015-01-01: 15_000
-JOINT:
-  2015-01-01: 25_000
-SURVIVING_SPOUSE:
-  2015-01-01: 15_000
-SEPARATE:
-  2015-01-01: 15_000
