--- policyengine_us/parameters/gov/contrib/co_general_assembly/family_affordability_credit/reduction/younger/amount.yaml
@@ -1,10 +1,0 @@
-description: The Colorado General Assembly reduces the family affordability tax credit by this amount for younger children.
-metadata:
-  period: year
-  unit: currency-USD
-  label: Colorado General Assembly family affordability tax credit younger reduction amount
-  reference:
-    - title: Colorado General Assembly House Bill 24-1311 Family Affordability Tax Credit
-      href: https://leg.colorado.gov/bills/hb24-1311
-values:
-  2024-01-01: 220
