+++ policyengine_us/parameters/gov/states/co/tax/income/credits/family_affordability_credit/amount.yaml
@@ -1,0 +1,10 @@
+description: The Colorado family affordability tax credit starts at this amount.
+values:
+  2024-01-01: 3_200
+metadata:
+  period: year
+  unit: currency-USD
+  label: Colorado family affordability tax credit amount
+  reference:
+    - title: Colorado General Assembly House Bill 24-1311 Family Affordability Tax Credit
+      href: https://leg.colorado.gov/bills/hb24-1311
