+++ policyengine_us/parameters/gov/states/co/tax/income/credits/family_affordability_credit/income_reduction/amount.yaml
@@ -1,0 +1,10 @@
+description: The Colorado family affordability tax credit is reduced by this amount as income increases.
+metadata:
+  period: year
+  unit: /1
+  label: Colorado family affordability tax credit income-based reduction amount
+  reference:
+    - title: Colorado General Assembly House Bill 24-1311 Family Affordability Tax Credit
+      href: https://leg.colorado.gov/bills/hb24-1311
+values:
+  2024-01-01: 6.875
