+++ policyengine_us/parameters/gov/states/co/tax/income/credits/family_affordability_credit/income_reduction/increment.yaml
@@ -1,0 +1,10 @@
+description: The Colorado family affordability tax credit is reduced based on these increments of adjusted gross income.
+metadata:
+  period: year
+  unit: currency-USD
+  label: Colorado family affordability tax credit income-based reduction increment
+  reference:
+    - title: Colorado General Assembly House Bill 24-1311 Family Affordability Tax Credit
+      href: https://leg.colorado.gov/bills/hb24-1311
+values:
+  2024-01-01: 5_000
