+++ policyengine_us/parameters/gov/states/co/tax/income/credits/family_affordability_credit/income_reduction/start.yaml
@@ -1,8 +1,8 @@
-description: The Colorado General Assembly reduces the family affordability tax credit starting at this adjusted gross income threshold, for younger children, based on filing status.
+description: The Colorado family affordability tax credit is reduced starting at this adjusted gross income threshold, based on filing status.
 metadata:
   period: year
   unit: currency-USD
-  label: Colorado General Assembly family affordability tax credit younger reduction start
+  label: Colorado family affordability tax credit income-based reduction start
   breakdown:
     - filing_status
   reference:
