+++ policyengine_us/parameters/gov/states/co/tax/income/credits/refundable.yaml
@@ -33,3 +33,11 @@ values:
     - co_low_income_cdcc
     - co_eitc
     - co_sales_tax_refund
+
+  2024-01-01:
+    - co_ctc
+    - co_cdcc
+    - co_family_affordability_credit
+    - co_low_income_cdcc
+    - co_eitc
+    - co_sales_tax_refund
