--- policyengine_us/reforms/co_general_assembly/__init__.py
@@ -1,3 +1,0 @@
-from .co_family_affordability_credit import (
-    create_co_family_affordability_credit_reform,
-)
