+++ policyengine_us/reforms/reforms.py
@@ -18,7 +18,6 @@ from .biden.budget_2025 import (
     create_medicare_and_investment_tax_increase_reform,
 )
 from .biden.budget_2025 import create_capital_gains_tax_increase_reform
-from .co_general_assembly import create_co_family_affordability_credit_reform
 from policyengine_core.reforms import Reform
 import warnings
 
@@ -55,9 +54,6 @@ def create_structural_reforms_from_parameters(parameters, period):
     capital_gains_tax_increase = create_capital_gains_tax_increase_reform(
         parameters, period
     )
-    co_family_affordability_credit = (
-        create_co_family_affordability_credit_reform(parameters, period)
-    )
 
     reforms = [
         afa_reform,
@@ -72,7 +68,6 @@ def create_structural_reforms_from_parameters(parameters, period):
         reported_state_income_tax,
         medicare_and_investment_tax_increase,
         capital_gains_tax_increase,
-        co_family_affordability_credit,
     ]
     reforms = tuple(filter(lambda x: x is not None, reforms))
