+++ policyengine_us/tests/policy/baseline/gov/states/co/tax/income/credits/family_affordability_credit/co_family_affordability_credit.yaml
@@ -1,9 +1,6 @@
 - name: Family with one kid under the age of 6
   period: 2024
-  reforms: policyengine_us.reforms.co_general_assembly.co_family_affordability_credit.co_family_affordability_credit
   input:
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[0].amount: 3_200
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[1].amount: 2_200
     people:
       head:
         age: 30
@@ -31,10 +28,7 @@
 
 - name: Family with one kid under the age of 6 and one under 16
   period: 2024
-  reforms: policyengine_us.reforms.co_general_assembly.co_family_affordability_credit.co_family_affordability_credit
   input:
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[0].amount: 3_200
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[1].amount: 2_200
     people:
       head:
         age: 30
@@ -60,15 +54,12 @@
         members: [head, dependent, dependent2]
         state_code: CO
   output:
-    co_family_affordability_credit: [0, 3_200, 2_200]
-    co_refundable_credits: 5_400
+    co_family_affordability_credit: [0, 3_200, 2_400]
+    co_refundable_credits: 5_600
 
 - name: Family with one kid under the age of 6 and one under 16, with reduction
   period: 2024
-  reforms: policyengine_us.reforms.co_general_assembly.co_family_affordability_credit.co_family_affordability_credit
   input:
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[0].amount: 3_200
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[1].amount: 2_200
     people:
       head:
         age: 30
@@ -94,15 +85,12 @@
         members: [head, dependent, dependent2]
         state_code: CO
   output:
-    co_family_affordability_credit: [0, 2_760, 1_870]
-    co_refundable_credits: 4_630
+    co_family_affordability_credit: [0, 2_760, 2_070]
+    co_refundable_credits: 4_830
 
 - name: Family with one kid under the age of 6 and one under 16, with reduction, fully phased-out
   period: 2024
-  reforms: policyengine_us.reforms.co_general_assembly.co_family_affordability_credit.co_family_affordability_credit
   input:
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[0].amount: 3_200
-    gov.contrib.co_general_assembly.family_affordability_credit.amount[1].amount: 2_200
     people:
       head:
         age: 30
